requirements.txt
Flask
Flask-Cors
rembg
Pillow
gunicorn